requirements.txt
aiohttp
aiosignal
asgiref
async-timeout
attrs
certifi
cffi
Django
django-cors-headers
django-filter
djangorestframework
exceptiongroup
frozenlist
h11
idna
Markdown
multidict
outcome
pillow
pycparser
PySocks
pytz
selenium
sniffio
sortedcontainers
sqlparse
trio
trio-websocket
typing_extensions
tzdata
urllib3
wsproto
yarl
